move data models out of types
